requirements.txt
Flask==3.0.3
pytest==7.1.2
pymongo==4.6.0
python-dotenv==1.0.0
certifi==2023.11.17